requirements.txt
requests
flask
python-dotenv
psycopg2-binary
pytest
mock
